+++ project.ewp
@@ -2042,9 +2042,15 @@
         <name>$PROJ_DIR$\sam3x8e.h</name>
     </file>
     <file>
+        <name>$PROJ_DIR$\sam3x8e_register_definitions.h</name>
+    </file>
+    <file>
         <name>$PROJ_DIR$\sam3x8e_setup.cpp</name>
     </file>
     <file>
         <name>$PROJ_DIR$\sam3x8e_setup.h</name>
+    </file>
+    <file>
+        <name>$PROJ_DIR$\sam3x8e_setup_definitions.h</name>
     </file>
 </project>
