+++ project.ewt
@@ -2326,9 +2326,15 @@
         <name>$PROJ_DIR$\sam3x8e.h</name>
     </file>
     <file>
+        <name>$PROJ_DIR$\sam3x8e_register_definitions.h</name>
+    </file>
+    <file>
         <name>$PROJ_DIR$\sam3x8e_setup.cpp</name>
     </file>
     <file>
         <name>$PROJ_DIR$\sam3x8e_setup.h</name>
+    </file>
+    <file>
+        <name>$PROJ_DIR$\sam3x8e_setup_definitions.h</name>
     </file>
 </project>
