+++ sam3x8e.cpp
@@ -6,157 +6,6 @@
 #include "intrinsics.h"
 
 /*
-    Enables the external crystal oscillator.
-*/
-void SAM3X8EClass::enable_main_oscillator(void)
-{
-    PMC_WPMR = WPKEY | WPEN_0;
-    CKGR_MOR = KEY | MOSCXTEN | MOSCXTBY | MOSCRCEN | MOSCRCF | MOSCXTST 
-                   | MOSCSEL | CFDEN;
-    PMC_WPMR = WPKEY | WPEN_1;
-}
-
-/*
-    Waiting for the external crystal oscillator to lock into it's
-    new frequency.
-    1: timeout
-    0: OK
-*/
-uint8_t SAM3X8EClass::wait_for_main_oscillator(void)
-{
-    uint16_t i = 0;
-    
-    while(MOSCXTS != (PMC_SR & MOSCXTS))
-    {
-        if(i++ > GLOBAL_TIMEOUT)
-        {
-            return 1;
-        }
-    }
-    return 0;
-}
-
-/*
-    Multiplies external clock frequency times 14.
-    12MHz * 14 = 168MHz
-*/
-void SAM3X8EClass::set_pll_and_divider(void)
-{
-    PMC_WPMR = WPKEY | WPEN_0;
-    CKGR_PLLAR = DIVA | PLLACOUNT | MULA | ONE;
-    PMC_WPMR = WPKEY | WPEN_1;
-}
-
-/*
-    Waiting for the pll to lock into it's new frequency.
-    1: timeout
-    0: OK
-*/
-uint8_t SAM3X8EClass::wait_for_pll(void)
-{
-    uint16_t i = 0;
-    
-    while(LOCKA != (PMC_SR & LOCKA))
-    {
-        if(i++ > GLOBAL_TIMEOUT)
-        {
-            return 1;
-        }
-    }
-    return 0;
-}
-
-/*
-    Prescaler divides the pll by factor 2.
-    168MHz / 2 = 84MHz
-    CSS selects this new clock as master clock.
-    1: timeout occured.
-    0: OK
-*/
-uint8_t SAM3X8EClass::select_master_clock(void)
-{
-    PMC_WPMR = WPKEY | WPEN_0;
-    PMC_MCKR = (PMC_MCKR & ~PRES_MSK) | PRES;
-    if(wait_for_master_clock())
-    {
-        return 1;
-    }
-    PMC_MCKR = (PMC_MCKR & ~CSS_MSK) | CSS;
-    if(wait_for_master_clock())
-    {
-        return 1;
-    }
-    PMC_WPMR = WPKEY | WPEN_1;
-    return 0;
-}
-
-/*
-    Waiting for the master clock to lock into it's new frequency.
-    1: timeout
-    0: OK
-*/
-uint8_t SAM3X8EClass::wait_for_master_clock(void)
-{
-    uint16_t i = 0;
-  
-    while(MCKRDY != (PMC_SR & MCKRDY))
-    {
-        if(i++ > GLOBAL_TIMEOUT)
-        {
-            return 1;
-        }
-    }
-    return 0;
-}
-
-/*
-    Reads the main oscillator frequency and stores the result in
-    main_oscillator_frequency.
-*/
-void SAM3X8EClass::check_main_oscillator_frequency(void)
-{
-    while(MAINFRDY != (CKGR_MCFR & MAINFRDY)){}  // This should be ready by now
-    main_oscillator_frequency = (((CKGR_MCFR & MAINF_MSK) / 16) * 32000);
-}
-
-/*
-    This is needed because the new clock frequency might be too high for the
-    flash.
-*/
-void SAM3X8EClass::change_flash_read_waitstate(int waitstate)
-{
-    EEFC_FMR0 = (waitstate << 8);
-    EEFC_FMR1 = (waitstate << 8);
-}
-
-/*
-    Sets the main clock frequency to 84MHz.
-    1: NOK (timeout)
-    0: OK
-*/
-uint8_t SAM3X8EClass::init_clock(void)
-{
-    uint8_t i = 0;
-    
-    change_flash_read_waitstate(5);
-    enable_main_oscillator();
-    i |= (wait_for_main_oscillator() << 0);
-    set_pll_and_divider();
-    i |= (wait_for_pll() << 1);
-    i |= (select_master_clock() << 2);
-    check_main_oscillator_frequency();
-    
-    if(i > 0)
-    {
-        return 1;
-    }
-    else
-    {
-        return 0;
-    }
-}
-
-/*
     Enables the peripheral clock for TWI1.
 */
 void SAM3X8EClass::init_twi1_clk(void)
