+++ sam3x8e_setup_definitions.h
@@ -1,0 +1,16 @@
+/*
+  sam3x8e_setup_definitions.h
+*/
+
+#ifndef SAM3X8E_SETUP_DEFINITIONS_H
+#define SAM3X8E_SETUP_DEFINITIONS_H
+
+#define DMA_CH                          0
+#define DMA_BUF_SIZE                    32
+#define DMAC_INTERRUPT_PRIORITY         (0)
+#define SSC_INTERRUPT_PRIORITY          (0)
+#define DIG_INP_INTERRUPT_PRIORITY      (5)
+
+#define GLOBAL_TIMEOUT                  0xFFDCU
+
+#endif
